backend/requirements.txt
Flask
PyPDF2
llama-cpp-python
huggingface_hub[cli]
